pyproject.toml
[tool.poetry]
name = "openvpc"
version = "0.1.0"
description = "open vpc service"
authors = [
    "Roberto Berto <roberto.berto@gmail.com>", 
    "Fabrício Radtke <fabriciossj@gmail.com>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
pydantic-settings = "^2.2.1"
uvicorn = "^0.27.1"
tortoise-orm = "^0.20.0"
aerich = "^0.7.2"
aiomysql = "^0.2.0"
httpx = "^0.27.0"
loguru = "^0.7.2"
pyvyos = {path = "../pyvyos"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.2"
coverage = "^7.4.4"
pytest = "^8.0.2"
faker = "^23.3.0"

[tool.aerich]
tortoise_orm = "src.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.pytest.ini_options]
testpaths = ["./tests"]
filterwarnings = [
    "ignore::DeprecationWarning:fastapi.*:", 
    "ignore::DeprecationWarning:tortoise.*:",
    "ignore::DeprecationWarning:httpx.*:"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
